mkdtimg: Use upstream python rewrite
The old C version is now deprecated, so we don't have to bother with
integrating that into the build system. That makes it quite easy for us
to bundle upstream libufdt as a submodule in dtc/libufdt (just beside
dtc/libfdt), and therefore not require exotic out of tree build
dependencies.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Julian Liu <wlootlxt123@gmail.com>
Signed-off-by: LibXZR <xzr467706992@163.com>
[Adjust for subtree]
Signed-off-by: Ratoriku <a1063021545@gmail.com>